Implemented Redux and migrated UserContext to Redux Store
