+++ package.json
@@ -9,8 +9,11 @@
     "firebase": "^9.9.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "redux-logger": "^3.0.6",
     "sass": "^1.54.3",
     "styled-components": "^5.3.5",
     "web-vitals": "^2.1.4"
