+++ src/App.js
@@ -1,12 +1,31 @@
 import { Routes, Route } from "react-router-dom";
+import { useEffect } from "react";
+import { useDispatch } from "react-redux";
 
 import Navigation from "../src/routes/navigation/navigation.component";
 import Home from "../src/routes/home/home.component";
 import Authentication from "../src/routes/authentication/authentication.component";
 import Shop from "../src/routes/shop/shop.component";
 import Checkout from "../src/routes/checkout/checkout.component";
+import { setCurrentUser } from "./store/user/user.action";
+import {
+  onAuthStateChangedListener,
+  createUserDocumentFromAuth,
+} from "./utils/firebase/firebase.utils";
 
 const App = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChangedListener((user) => {
+      if (user) {
+        createUserDocumentFromAuth(user);
+      }
+      dispatch(setCurrentUser(user));
+    });
+    return unsubscribe;
+  }, [dispatch]);
+
   return (
     <div>
       <Routes>
