+++ src/components/directory-item/directory-item.styles.scss
@@ -5,7 +5,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  border: 2px solid #D9BE83;
+  border: 2px solid #d9be83;
   border-radius: 25px;
   margin: 0 7.5px 15px;
   overflow: hidden;
@@ -15,16 +15,13 @@
     transform: scale(1.02);
     transition: transform 0.5s cubic-bezier(0.25, 0.45, 0.45, 0.95);
 
-
-
-    // & .background-image {
+    // & .backgroundImage {
     //   transform: scale(1.1);
     //   transition: transform 6s cubic-bezier(0.25, 0.45, 0.45, 0.95);
     // }
 
     & .body {
       opacity: 0.9;
-
     }
   }
 
@@ -40,7 +37,7 @@
     margin-left: 7.5px;
   }
 
-  .background-image {
+  .backgroundImage {
     width: 100%;
     height: 100%;
     background-size: cover;
@@ -54,7 +51,7 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    border: 1px solid #D9BE83;
+    border: 1px solid #d9be83;
     border-radius: 25px;
     background-color: white;
     opacity: 0.9;
@@ -67,22 +64,19 @@
       p {
         color: white;
       }
-
     }
 
     h2 {
       margin: 0 0px 0;
       font-size: 25px;
-      color: #D9BE83;
-    } 
+      color: #d9be83;
+    }
 
     p {
-      font-weight:bold;
+      font-weight: bold;
       font-size: 20px;
       margin: 0;
       padding: 0;
-      
-
     }
   }
 }
