+++ src/store/root-reducer.js
@@ -1,0 +1,6 @@
+import { combineReducers } from "redux";
+import { userReducer } from "./user/user.reducer";
+
+export const rootReducer = combineReducers({
+  user: userReducer,
+});
