+++ src/store/user/user.types.js
@@ -1,0 +1,5 @@
+export const USER_ACTION_TYPES = {
+  SET_CURRENT_USER: "SET_CURRENT_USER:",
+};
+
+export default USER_ACTION_TYPES;
